[fdf][compat] Support fragment names
Make it simple to support fragment names in the compat shim.
We can figure out the fragment names based on the device service
instances that are given to us.

Unfortunately, we have to add some node information to the DFv1
CML file so the instances are routed. Add a GN argument for this
so it can happen easily (and so it can be easily replaced when
we move to services).

Bug: 93678

Change-Id: I607371eea1849b7737c5cea54cec37da208c8bec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674084
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>