[pkg-cache] BasePackages uses HashMap for package hashes
Enables O(1) checking of if a package is in base. Useful for after
the pkgfs migration when this check will be on the hot path.

fuchsia.pkg/PackageCache.BasePackageIndex now returns entries
sorted by url. The ordering used to be the same order as the
static_packages manifest with the system_image package appended.

Change-Id: I09d2ae22a4af62bc81ca1e5775fd446207b7938c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672876
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>