[wlan][sme] Consider driver support when configuring protection.
This change considers driver support for WPA3 when constructing an
`Authentication` from an SME `ConnectRequest`. This more closely mirrors
the original code this was based on and effectively allows WPA2 to be
used with transitional networks when the client does not support WPA3.

Bug: 99317
Testing: Unit tests and manual testing.
Change-Id: I5b8f4f884ab6fc8081d9824e427a7ca0e7affb96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673164
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Reviewed-by: Yung-chieh Lo <yjlou@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>