[ermine] Fix for screen flicker with overlays on top of child views
The screen flickers when overlays are visible on top of child views.
Ermine draws a semi-transparent scrim that darkens the content of the
child view in order to make the overlays stand out. But for some reason
it has causes flickering.

This is being investigated at the Graphics layers. Until then we are
removing the semi-transparent scrim.

Bug: 99036
Change-Id: I08b7107da84b7f7f81badf2cd841fdafab1f81cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/674608
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>