[cm] Store fasync::Task instead of an Abortable for dir-waiter
Directory waiter futures are wrapped in an abortable before
running on a detached fasync::Task.

The abortable is stored in IncomingNamespace and is aborted
explicitly when IncomingNamespace is dropped.

When a task is dropped, its future is no longer polled by the
executor.

Instead of using an abortable, run the waiter in an fasync::Task
directly and store that task in the IncomingNamespace.

Test: fx test component_manager_tests
Change-Id: I95ec6dd982f1bce907ea54456fbab2e7ab4d8034
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674966
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>