[fdf][ramdisk] Add wait_for_device to biotime test
The biotime test occasionally flakes because the
test can run before the ramdisk is made available,
causing the ramdisk_create to fail.  This CL adds
a wait to the setup of the test to make sure that
the ramdisk is available.

Testing: this is a change to a test.

Change-Id: Ife8f602fb596829c69a2ced45db3eb92cabc48e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674702
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>