[inspect] Add AtomicUpdate to C++ library
Add Node::AtomicUpdate(), which allows the caller to perform multiple
operations that are guaranteed to appear in the same generation.

Introduces State::{Begin,End}Transaction() to support this new
mechanism. While one or more transactions are open on a given State, all
operations will appear in the same generation -- regardless of the
thread from which they are performed. This is aligned with the Rust
library semantics.

Bug: 97904
Change-Id: I68db6cad2ee550862615dd16661ab9544ba177b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670768
Commit-Queue: Christopher Masone <cmasone@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>