[blackout] use ffx to run commands on the target
Set up an isolated ffx instance to communicate with the target device
and run the target components. This removes the need to call ssh on our
own, or do target discovery.

As part of this change, we swap the meanings of some return codes,
making verification failures need a specific return code and retrying on
any other errors, making the retry logic more resilient.

This also adds an integration test for blackout that runs in CQ,
providing a backstop for the host-target-interaction functionality.

Bug: 44392
Bug: 96569
Test: ffx storage blackout blobfs_checkerboard
Test: ffx storage blackout integration
Test: fx test --e2e blackout-integration
Change-Id: I7d373ce4693b797c6e48c4fba0f911d9bf9cab79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662499
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>