[cpp][style] Guidance for naming generated impl headers.
This has come up as a topic in the RFC for configuration accessors
which need to generate these headers. They will not go into an
SDK, and are generally "private" to the component using them.
The current guidance for generated headers assumes they are part
of the platform, and only provides guidance for `fidl/`, which
does not (currently) describe structured config libraries.

Bug: 95371
Change-Id: I38f0dc80772df5e70efe980c2acb1bee850fd336
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673667
Commit-Queue: Adam Perry <adamperry@google.com>
Reviewed-by: Adam Barth <abarth@google.com>