[media/vnext] add buffer lead time tracking to stream_io lib
This CL adds lead time tracking to the stream_io library. Lead time
tracking assigns a lead time to newly-allocated buffers, indicating
how far in advance of allocation the release of buffers is occurring.
Given that buffers may be allocated out-of-order with respect to when
they were freed, the arrival time of a buffer (the time at which it
is freed) is pushed into a queue. When a buffer is allocated, the
current time is compared to the arrival time at the head of the queue
to determine lead time.

Test: unit test added
Change-Id: Ibf942a60f61ab8cb365e4f73d8b54ea245b5e018
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671125
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>