[build][assembly] Begin splitting base_package_labels into developer_ and legacy_
This is the first step in a soft transition.

This creates a new set of GN args ('legacy_*') that mirror the
'*_package_labels' set of args. These are then used to create two new
sets of GN target groups:

'//:legacy_base_package_labels'
and
'//:developer_base_package_labels'

which are used to contain base package labels set by the product
definitions and the developer (via args.gn), respectively.

This first CL establishes the new args and groups, but doesn't remove
the existing, due to transitioning with the vendor repos.

Bug: 99025
Change-Id: I252dd4b53813b61eaab0375432b43e50e83c31b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671796
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>