[fxfs] [recovery] Overwrite Fxfs superblocks as part of FDR.
Support overwriting of Fxfs superblocks so that we can perform factory
reset of devices without zxcrypt.

This is only step one.
Following this, we need to:

 1. Move factory reset to fshost. We can then shutdown the filesystem
    before overwriting, removing a race that would prevent reformat.
 2. Implement keybag deletion under Fxfs.

We will not be in a shippable state until both these things are done but
will benefit in the interim in being able to unblock zxcrypt removal for
Fxfs native-ecrypted volumes.

Bug: 98889
Change-Id: I8416a07840add18d97fe54e5b065f8b2419bb824
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674325
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>