[fdf][storage] Change factory_reset_test to use ramdisk
The factory reset test uses a function in the fs_management
that will in the future assume that the file descriptor argument
refers to a block device.  Rather than alter that function
to handle both vmos and block devices, we alter this test to
have it use a block device.

Bug: 91608
Test: this is a change to test code.
Change-Id: I0d8766fe14b7f0d71c25435aceca14346aab70e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674584
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Garratt Gallagher <garratt@google.com>