[codec][vaapi] Implement diagnostics for VAAPI
Implemented the diagnostic data in a way that can both be used by
amlogic driver and the vaapi driver. A function was added to
CodecAdapter that allows the driver to provide the new created codec
with a reference so that it can create codec diagnostic information.
The codec is agnostic to how the diagnostic information was created
since, as of now, amlogic is a driver and vaapi is a component.

Bug: 81304

Change-Id: Ib80804682aa54c1e9b4b4b3141489aa622563751
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672564
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>
Reviewed-by: John Bauman <jbauman@google.com>