[ffx][test] Retain and save all suites at once
This updates the directory reporter to retain all results in memory,
and save all results when either a suite or test run completes. This
essentially removes an optimization where suites were saved and evicted
from memory as soon as they completed.
While this might regress performance somewhat, we will need to do this
anyway when we transition to a single file format in the stable test
output. This change is in a separate commit to ease reverting in case
this causes issues.

This change also removes some unit tests related to the optimization
that are now obsolete.

Testing: run CQ in the *ffx builders
Existing tests

Bug: 81195

Change-Id: I303c1543512460e65e27dab38f5ad6ce22f94f0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674748
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>