[ffx] Add ffx profile memory.
This CL allows ffx to print a summary of how the memory is used on the
remote device.

A new FIDL api is added to allow ffx to send a socket handle to
memory_monitor.

ffx creates a socket and sends one of the two handles to memory_monitor.
memory_monitor writes all of its data to the socket.
At the same time, ffx reads all the data sent to the socket,
and then prints it out.

Test: fx test ffx_profile_memory_lib_test
Change-Id: I49c4109228f455022b42ee189ea00a3b99c2e2fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670542
Reviewed-by: Dave Moore <davemoore@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Jean-François Geyelin <jif@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Pierre Labatut <plabatut@google.com>