[virtualization] Relax header size check
We were enforcing that the disk size must be aligned to the cluster
size. This is not something required by qcow. The block device already
has the sufficient checks in place to reject any request that extends
beyond the device capacity so this check can be removed.

Bug: 95529
Test: Booted debian guest (required copy-on-write patches). Fixes an
      error when booting due to triggering the removed error
      condition.
Change-Id: I1fb97caeccfd370422eae4dea96e616bb5520a7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675907
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: David Hastings <dahastin@google.com>