[legacy-boot-shim] Use UartItem
This is because the ACPI tables are now parsed on InitMemory,
and gLegacyBoot.acpi_rsdp && gLegacyBoot.uart are set at that point.

Move LegacyBootShim to use boot_shim::UartItem, instead of AcpiUartItem,
now that the ACPI tables are parsed and the relevant information is
within LegacyBoot.

Removed acpi.cc/.h and updated tests to verify that the boot shims
propagates the values from LegacyBoot.

Bug: 97896
Test: fx test -v -o legacy-boot-shim-tests
Change-Id: Ic6eab2ed8a7ba572a3137ef8c8e7b05f55babef8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666942
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>