[fake-clock] fix link args
Shared libraries get put in the root_out_dir by default.

It was working before because the relevant path happened to already be
on the linker search path everywhere this was used, but we shouldn't
rely on that.

Change-Id: Iffba6052e606ed343ca362a5e75c103e3b2873ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675910
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Dan Johnson <computerdruid@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Fuchsia-Auto-Submit: Dan Johnson <computerdruid@google.com>