[a11y] Fix flake in flatland-a11y-view-tests
This test constructs a flatland display, and attaches the a11y view and
a child view of the a11y view ("the proxy view"). The test fixture
determines that the a11y view has been inserted by waiting until the
proxy view is connected to the display. Prior to this change, the test
only polled the proxy view's parent viewport status once. In most cases,
the first status reported was CONNECTED_TO_DISPLAY, and the test passed.
However, it was possible for flatland to report DISCONNECTED_FROM_DISPLAY,
as the first status. In this case, the test would never call
ParentViewportWatcher::GetStatus() again, and would hang.

This change ensures that the test fixture continues to poll the proxy
viewport watcher until the proxy view is connected.

Test: flatland-a11y-view-test
MULTIPLY: flatland-a11y-view-test: 200
Bug: 99629
Change-Id: I5b61ed22b438b24ac5839dd4e9b70274dfc0a971
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676323
Commit-Queue: Alexander Brusher <abrusher@google.com>
Reviewed-by: Catherine Duncan <catduncan@google.com>