[check-licenses] Support explicitly adding licenses to the NOTICE file.
We currently use the output of "gn desc" to filter the full NOTICE down
to only include license files that actually contributed to the build.
This works well, but there are some situations when we want to add additional license texts to the NOTICE file, e.g. b/229106823.

This CL modifies check-licenses to take a list of "filter"s, which is a list of string paths to JSON files. The JSON file format is a simple string list, where each entry is a path to a license file or project. Check-licenses will read through these files, and include every associated project in the final NOTICE file.

Bug: b/229106823

Change-Id: I45bff7aa0f526dd4fc2b5ec41c19ffaecfe50961
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675057
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>