[testrunner] Create stdio file even if empty
Currently, if a test doesn't produce any stdio then testrunner will skip
writing a stdout-and-stderr.txt file, but still declare in summary.json
that such a file exists. This causes failures when the recipe tries to
read the file that doesn't exist, e.g.
ci.chromium.org/b/8815013275155244257.

For simplicity, we now always write the file even if it's empty. I also
considered not adding the file to summary.json if it was empty, but that
made testrunner's unit tests more complicated because the mock data
needed to conditionally include stdout-and-stderr.txt depending on
whether the mock output was empty.

Change-Id: I66b399f20ec8fab8744a7a7722318bebc38d9e56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676783
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>