[kazoo] Print instructions before the diff on fail
In the kazoo golden test it would previously print the diff followed by
instructions on how to update the golden file. However, if there are
many failures the output will get truncated by the bots and the
instructions will be lost.

This prints the instructions first so they will not be lost in buildbot
output.

Change-Id: I351fae81cf077e8e1c1e1dc3e6d3f5cc368ae08c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676842
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>