[codec] Replace system_clock with steady_clock
In order to wait for a set amount of time we should use steady_clock instead
of system_clock. steady_clock is guaranteed to be a monotonic clock. Currently
we were using system_clock which is only guaranteed to be monotonic iff
is_steady is true which is platform dependent.

Fixed: 99758

Change-Id: Iccbd7817db4d04340ce779da7aaa881c1d603730
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675345
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>