[fxfs] Track allocations by owner_object_id.
This is a breaking change (Fxfs version 15->16).

The owner is the ObjectStore that holds a file so
this change tracks the bytes allocated to each ObjectStore.

This information will be used when deleting encrypted ObjectStore
as well as in enforcing storage quota limits.

Note that the space consumed by a volume is more than just
the bytes allocated to owner_object_id for the volume's ObjectStore.
It should also include the StoreInfo file and layer files (stored in
the Root ObjectStore).

Change-Id: I003746305d30f7bd0592d28215f65e259116f01b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675328
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>