[ermine] Fix gray screen when app does not load from package server
This change fixes an issue when an app is launched, in this case,
spinning_square_view, but fx serve is not running. This results in a
gray screen and the user cannot bring the system overlays back.

The fix is to allow user to invoke system overlays even while waiting
for the application to launch. This works because the shell still has
focus. And when the app does launch, the system overlay is dismissed
when the app receives focus.

Fixed: 99699
Change-Id: I89c5f1cce2ac6633f28843b159c462d9fd42921a
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/677006
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>