[fxfs] Add support for volume creation/enumeration
Implements the Volumes API on Fxfs.

Also adds the volumes directory, which relies on
vfs::directory::immutable::Simple and as such inherits most directory
functionality for free.  Note that this isn't wired up into the export
root yet, that needs to be done in tandem with fs_management changes to
adjust how Fxfs is launched.

Also adjusts the naming of a few methods in the API to match the RFC,
and moves removal to be within Volumes so that we can internally rely on
the VFS' Simple but still retain control over how removal works.

Bug: 99182
Change-Id: I1ab945a5f9c0cdaeb51cab709e56878482ccedf6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676342
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>