[overnet][fasync] Cancel redundant overnet port waits
Overnet signal proxying creates a lot of OnSignals objects, each of
which calls zx_object_wait_async on creation. But if the handle in
question is never closed, and the event never happens, the port
registration sticks around even after the OnSignals is dropped and the
packet would thus not have anything to handle it. Eventually the linked
list of port registrations gets too big and the kernel kills
overnetstack.

This CL adds a Drop impl to OnSignals which cancels the port wait. This
requires holding a reference to the handle inside of OnSignals, which
requires adding a lot of explicit lifetimes in overnet to prove that the
handle isn't closed while the OnSignals exists.

Fixed: 95247
Fixed: 96326

Change-Id: I005b0bbd76d985a2261cd19112bed9b5676a8cef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675183
Commit-Queue: Theodore Dubois <tbodt@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>