[settings] Move from //garnet to //src
This change moves the source code of the settings service from //garnet
to //src as //garnet is deprecated. Proxy targets are left behind in
//garnet so that existing workflows and references don't break.

Some files in the meta folder are not being migrated in this CL due to
dependencies on the actual file path, which can't be solved with proxy
targets. Those will be untangled separately.

Fixed: 98248
Test: CQ, verify that developer workflows still work locally
Change-Id: Ib81eb451db391b0a37bb914ddf071de28686a020
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676811
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Sijie Chen <sijiec@google.com>