[text_manager] Use text_manager as CFv2 component
text_manager was migrated to CFv2. This change removes CFv1 from
use.

text_manager is placed in /core for the time being, up until
/core/ui is ready to accept it.

The CFv2 component is added to the same package as the CFv1. Since the
binaries are identical, the change in package size should be negligible.

Bug: 90441
Change-Id: Icc284c33fdede499ee0d0e60ba59f9bf8cea439a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674182
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>