[docs][sdk] Clarify git/bazel install.
Previously the "Install dependencies" section started with a bunch of
text and links to the project documentation pages for Git and Bazel. The
download links were hidden inside a list after some more text and a
"Note" box and I completely missed them the first time through.

This removes the introductory text (it's mostly duplicative of the
existing links) and moves the "Note" to below the download list. The
links are moved to the beginning of each list item so it's more clear
what to do.

Added some explanatory text on the Bazel item which links directly to
downloading the Bazelisk binary. The Bazel web site recommends this
approach but getting the install links require several decisions and
clicks. This also clarifies that the Bazelisk binary behaves as the
"bazel" binary (this detail is buried under the "official readme" of
Bazelisk and cost me several minutes of confusion).

Change-Id: I9c4b1f70d74685c797d4678e4f3f600cee786ac3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677165
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>