[fidl] Remove NaturalBodyEncoder::GetBodyView
Before this CL, a BodyView struct was created and then the values were
copied to an OutgoingMessage. After this CL, the OutgoingMessage is
directly created.

The original goal of this CL was to improve performance, but it
doesn't have much effect. Still, this CL does simplify things.

Change-Id: Iaed5fca1116de54c8b512fc26c36201694266d19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669550
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>