[codec][vaapi] Fix race allocating output buffers
ProcessOutput may be called before the sysmem buffers are allocated.
The code to determine the output buffer size could assert because it
depends on the output buffer constraints.

If we move that calculation code after
output_buffer_pool_.AllocateBuffers then it's guaranteed that the
buffers are allocated by that point.

Fixed: 99841
Change-Id: I5c8dff3faa52ba967238a81243e5707212ed6272
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677085
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Stefan Bossbaly <stefanbossbaly@google.com>