[netstack3] Move SLAAC to IP
This allows us to properly remove SLAAC timers when a SLAAC address is
manually removed and on interface disable (later changes).

Follow-up changes will refactor the SLAAC implementation and tests to
take advantage of the contextification introduced in this CL. This CL
does not perform the refactors as this change is already quite large
and a pure code move followed up by a refactor is easier to review
than a code move + large refactor.

Bug: 93817

Change-Id: Ibd36799ccec0f00b3d180af177d0b3bcc34bcda8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676244
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>