[Brightness manager] Auto-brightness in Brightness Manager.
This CL add a small, auto-brightness component to be used in basic
systems. It dims and brightens the screen and the LED(s) as
the light on the light sensor changes.

It is very simple and completely self-contained.
There are no controls or options for it.

It is a separate target (auto-brightness) in the build file and
it contains a shard that can be added to existing configurations
as required.

Starts automatically if included in a build.

Common files between Brightness manager and Auto-brightness have
been moved to a library.

Test: New tests added, all tests pass
Bug: b/226527434
Change-Id: Id4efdbe38748316e7b87557e375baaaa4164bf45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673306
Commit-Queue: Kevin Twidle <kpt@google.com>
Reviewed-by: Alex Wong <alexdwong@google.com>
Reviewed-by: Mike McCreavy <mccreavy@google.com>