[fshost] Make first boot quieter
When booting with Fxfs for the first time, the data partition is empty,
which will trip an fsck failure and log a scary message.

Change fshost to first check the content format; if it doesn't look
right, then just go straight to reformatting.  This might mean that some
legitimate corruptions (i.e. those that overwrite the first block)
aren't caught as true errors, but the medicine is probably worse than
the poison in this case.

Change-Id: I84f441b64cf927f59debb7dfe95018abc8eb0b3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675185
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>