[mediaplayer] remove reference to buffer_internal.h
mediaplayer used to reference ffmpeg's buffer_internal.h in order to
access a field in an AVBuffer. This is a problem,
because buffer_internal.h reference <stdatomic>, which is not
compatible with our C++ toolchain going forward.

Fortunately, there is a public method to access the field in
question, so the reference to buffer_internal.h may be removed.

Test: no behavior change

Fixed: 99837
Change-Id: I25b1a88e9bf4ade5aa327a76ac7c8e8469ffe592
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677802
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>