[camera] Restrict output frame rate to 24fps
This CL adds a rudimentary capability to limit the frame rate of the
camera platform. It starts with adjusting the sensor frame rate, which
automatically limits the rate of frames passing through the MIPI CSI
interface and the ISP. It also provides for dropping frames within the
post-processing pipeline to further reduce the frame rates. It does not
alter the stream descriptions used by clients to avoid introducing
potential problems with them.

Bug: b/229264734

Change-Id: I3cd0c1e58d3210b567ec22dff49f59691e291fef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669431
Commit-Queue: John Sasinowski <jsasinowski@google.com>
Reviewed-by: Ernest Hua <ernesthua@google.com>
Reviewed-by: David Rees <drees@google.com>