[flatland] Update ViewWatcher Logic
Update the link system to make sure that the viewport
properties being calculated and returned to clients are
those coming from the subtree in the transform graph with
the lowest overall scale factor. This only applies in cases
of a transform having multiple parents.

New unittest in flatland_unittests

Fixed: 99895

Change-Id: I4a6100ac69531e679f42eab26b1c2fad404a3b9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675222
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Chris Fontas <cfontas@google.com>