[ffx] Move upscript out of template
Previously, we made the decision to start the emulator upscript by
passing the script path to the qemu backend and allowing the backend to
invoke the script. This was contrary to how fvdl had been designed to
run the script before invoking the backend, because we wanted to allow
for customization with the flag template. However, since that design
decision, we've learned that we don't want users modifying the template
manually at all, and the upscript will be set up in the ffx config. It's
also led to issues like fxbug.dev/98874 and fxbug.dev/98880 due to bad
interactions with the shell and launcher process.

This CL moves the upscript path out of the emulator flags template, and
instead launches the upscript directly from the launcher before
launching the emulator itself. This makes for a cleaner interaction with
any sudo prompts, and ensures there's only one child process at a time
so that stdio is properly serialized.

Bug: 98874, 98880
Change-Id: If92404eb7fbd9ca3fae30061ed76e878975b334a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677807
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>