[kernel][socket] Reset pkt_len_ in MBuf
The pkt_len_ field is used by the Read code to determine if a given
MBuf is part of the datagram chain or not. Since pkt_len_ is not reset
when a MBuf is recycled, reading and writing different sized datagrams
could cause a datagram head packet (with a pkt_len_ set) to get reused
as a datagram body packet.

Change-Id: If1d178a748d1903f0bfccdd7791cdcf74af47222
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677864
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>