[assembly] Add fuchia legacy input bundle archive to uploads
Add //build/images/fuchsia:fuchsia_legacy_input_bundle to the set of bundles
that are documented in the build_api.

Threads a parameter on assemble_system() through to the creation of the
legacy bundle, and then only sets that flag on the main fuchsia assembly.

Also changes legacy.gni to create the package and archive using deps
instead of just creating the targets (but not depending on them).

Fixes: 100066
Change-Id: I8b1f024927d94cf0c313dca14156b3ec0da7907a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678296
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>