[tools] Fix VM RAM allocation in //tools/emulator
Remove support for VMs with <1MB of RAM. QEMU 6.2.0 (which we now use)
doesn't support values that low.

At some point, QEMU began to expect values provided for the
`-m` command-line switch to be an integral number of megabytes --
not bytes. Our code assumed the latter, which meant that the default
value of `8K` used in //tools/emulator got parsed to `8192` and
handed off to QEMU...which happily started a VM with 8GB of RAM.

Change-Id: I94029533bf8f37f37ef0dc04ed15a4ec0c8e3713
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678262
Commit-Queue: Christopher Masone <cmasone@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>