[physboot] Increase mexec hand-off scratch space
This space is currently limited to a page (seemed large enough at the
time). We are hitting this limit now, however: a particularly fragmented
EFI-level view of the physical address space is lending itself to a
MEM_CONFIG ZBI item that is just about a page in size itself, which we
then fail to account for in the scratch space.

A more robust fix is coming, but this is a good enough to address the
associated bug for now.

Bug: 99875
Change-Id: Ib5dc17918f6fffc05c4a8a8c8d58dbd3deaa4a84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678225
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Curtis Galloway <curtisgalloway@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>