[kernel][mmu] Expand consistency check for arm64 PTs
The purpose of this change is to expand our consistency checking to
help isolate a seemingly map-after-destroy issue we're seeing in the
field.

Add a new method, DisableUpdates, to ArchVmAspaceInterface that puts
the arch aspace into read-only mode and asserts that it contains no
mappings.

Call DisableUpdates as part of VmAspace::Destroy.

Update ArmArchVmAspace's methods that modify PTs to assert that
updates are enabled.

No functional change to x86.  DisableUpdates is a no-op on x86.

Bug: 79118
Change-Id: Ia88cdd981f91af034176e907fba0f2fd5d8e98f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678097
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>