[kernel] Pass process to HandleTable methods that enforce policy
Most HandleTable methods enforce job policy by terminating a process
that uses a bad handle value.  Currently, a HandleTable is owned by
exactly one process so it makes sense that HandleTable keeps a
reference to its owning process and simply calls the
EnforceBasicPolicy method on it when necessary.

In the future a HandleTable may be shared among several processes.  In
a world where a HandleTable may be shared, there is no single onwing
process.  We'll need a way for HandleTable to call EnforceBasicPolicy
on the proper target.  An obvious solution for HandleTable to use the
"ambient" process via static method ProcessDispatcher::GetCurrent.
This approach is somewhat sublte and could lead to surprises down the
line (what happens when a pure kernel thread, with no user mode
component, wants to call one of these methods that assume a current
process?).

This CL provides a different solution.  Instead of using a field
pointing to the owning process or using the ambient process,
HandleTable now requires callers to pass in a ProcessDispatcher& when
they call policy enforcing methods.

Why pass by reference?  Pass by reference to drive home the point that
it's non-optional and to make it clear that the callee is just getting
to "borrow" for the duration of the call.

Remove the unused HandleTable::IsHandleValid method.

Bug: 98922
Change-Id: I3cacf3b3f6fb957130e14fe741c498ef4b1a0ad0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677650
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>