[ffx] Add `target add` auto-wait.
This adds support for auto-wait. Auto-wait is off by default (to be
enabled later on).

The implementation for `wait` is functionally the
same as running `ffx target wait` immediately after `ffx target add`
without a wait invocation, though it happens internally.

The behavior of `ffx target add` is also changed so that if waiting, and
the added target eventually fails to connect, the target will be removed
and the client will receive the error caused during the connection
process. The target will also be removed if the client disconnects in
the middle of adding the target.

Bug: 89546

Change-Id: Ia382dd8bb63ec3ae8c9e4136021fce7021c95fc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671732
Reviewed-by: Casey Dahlin <sadmac@google.com>
API-Review: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>