[memory] Use the noexcept version of std::filesystem::exists
The old code would throw an exception if the monitored file
does not exist, causing the memory monitor to crash.

This does not seem necessary, so I opted to use the
overload that does not throw an exception, allowing the
program to print an informative message.

With this realization, it seems that this crash might
be a red herring to the reported bug.  Except if
memory monitor crashing pulls with it the web engine
as well, which IIRC requires it as a service.

Bug: 99414
Change-Id: I0a60f876c2903a8f4f2bc23c87de2ac37363037b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677704
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>