[credential_manager] Recovery Flow.
This CL introduces the state replay feature of the PinWeaver Protocol
into the Credential Manager implementation. When launching Credential
Manager we first validate that the root hash stored on disk is the
same as the root hash stored on the CR50 chip. If these values are
not the same we will retrieve a log from the CR50 of applied
operations. If we are exactly one step behind the CR50 we will
attempt to replay the one operation we are behind. This can
be one of four operations:

- InsertLeaf: We need to replay an InsertLeaf operation.
- RemoveLeaf: We need to replay a RemoveLeaf operation.
- ResetTree: We need to replay an entire ResetTree operation.
- TryAuth: This is a special case where we need to send a
  command (log_replay) to the CR50 and then update the hash tree
  and lookup table with the corresponding response data.

If the replay_state function succeeds we can continue on with
launching the system. This should only ever occur if we fail
to write out to disk after performing a mutable operation
on the CR50.

Bug: 96663
Change-Id: I38286672e5fee6e70f67c0e87411e15606508538
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673711
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Ben Wright <benwright@google.com>