[debugger] Get relative paths from line tables
This is the intended behavior since we always want to get the relative
path against the compilation directory.  It was not a problem because
for in-tree binaries, the compilation directory is always "." thus the
absolute path and the relative path is the same.  However, in Bazel SDK
the compilation directory is set to "/proc/self/cwd", which we have to
ignore and replace with the real build_dir from the symbol index.

Bug: 99751
Change-Id: Iee3958be636819fe7ae978af5ff4e1e8193437fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678509
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>