[fx] Remove dependency on python
fx self-tests currently rely on python if 'realpath' is not available,
for example on MacOS. However, python is not always available on our
trybots and should not be relied upon. This CL changes the
implementation to use a shell-based fallback that walks up the parents
of the given path until it finds a valid parent, then uses cd/pwd on it,
and concatenates with the rest of the given path.

Cq-Include-Trybots: luci.fuchsia.try:core.x64-host_test_only-mac
Bug: 99598
Change-Id: Iabc60a4af5694894b15ad7a31a185280c31cb2d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678645
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>